// openpectus/frontend/src/app/api/model/batchJob.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Represents a current or historical run of a process unit. 
 */
export interface BatchJob { 
    id: number;
    unit_id: number;
    unit_name: string;
    completed_date: string;
    contributors?: Array<string>;
}